Refactoring everything.
