+++ frontend/src/App.jsx
@@ -2,23 +2,19 @@ import "./App.css";
 import { Route, Routes } from "react-router-dom";
 
 //Component imports
-import Footer from "./Components/Footer";
+import Footer from "./Components/Footer/Footer";
 import Nav from "./Components/Nav/Nav";
 
 //Route imports
 import Home from "./Pages/Home";
 
 //Pages Routes
-// import DesignStart from "./Pages/WhiteLabeling/DesignStart"
-import PartnerWhiteLabel from "./Pages/WhiteLabeling/PartnerWhiteLabel";
-import RequestConfirmation from "./Pages/WhiteLabeling/RequestConfirmation";
-// import SiteBuildA from "./Pages/WhiteLabeling/SiteBuildA"
-import RequestSite from "./Pages/RequestSitePage/RequestSite";
-// import SiteBuildC from "./Pages/WhiteLabeling/SiteBuildC"
-import ReviewAndSubmit from "./Pages/WhiteLabeling/ReviewAndSubmit";
-import OptionalQuestions from "./Pages/OptionalQuestionsPage/OptionalQuestions";
-import CriticalQuestions from "./Pages/CriticalQuestionsPage/CriticalQuestions";
-import OptionalContactInfo from "./Pages/OptionalContactPage/OptionalContact";
+
+import PartnerWhiteLabel from "./Components/WhiteLabelSite/PartnerWhiteLabel/PartnerWhiteLabel";
+import RequestConfirmation from "./Components/WhiteLabelSite/RequestConfirmation/RequestConfirmation";
+import ReviewAndSubmit from "./Components/WhiteLabelSite/ReviewAndSubmit/ReviewAndSubmit";
+import WhiteLabelPage from "./Pages/WhiteLabelPage/WhiteLabelPage";
+import SiteDemo from "./Pages/SiteDemo/SiteDemo";
 import { FormProvider } from "./Context/FormContext";
 
 function App() {
@@ -28,14 +24,9 @@ function App() {
       <FormProvider>
         <Routes>
             <Route exact path="/" element={<Home />} />
-            {/* <Route path="/DesignStart" element={<DesignStart />} /> */}
             <Route path="/PartnerWhiteLabel" element={<PartnerWhiteLabel />} />
-            {/* <Route path="/SiteBuildA" element={<SiteBuildA />} /> */}
-            <Route path="/RequestSite" element={<RequestSite />} />
-            {/* <Route path="/SiteBuildC" element={<SiteBuildC />} /> */}
-            <Route path="/CriticalQuestions" element={<CriticalQuestions />} />
-            <Route path="/OptionalQuestions" element={<OptionalQuestions />} />
-            <Route path="/OptionalContact" element={<OptionalContactInfo />} />
+            <Route path="/WhiteLabelPage" element={<WhiteLabelPage />} />\
+            <Route path="/SiteDemo" element={<SiteDemo />} />\
             <Route path="/ReviewAndSubmit" element={<ReviewAndSubmit />} />
             <Route path="/RequestConfirmation" element={<RequestConfirmation />} />
         </Routes>
