--- frontend/src/Pages/OptionalQuestionsPage/OptionalQuestions.jsx
@@ -1,12 +1,0 @@
-import OptionalQuestions1 from "../../Components/OptionalQuestions/OptionalQuestions1"
-import OptionalQuestionsInfo from "../../Components/OptionalQuestions/OptionalQuestionsInfo"
-import "../../Components/OptionalQuestions/OptionalQuestions.css";
-
-export default function OptionalQuestions(props) {
-    return (
-    <div>
-        <OptionalQuestionsInfo/>
-        <OptionalQuestions1/>
-
-    </div>
-    )}
