--- frontend/src/Pages/WhiteLabeling/DesignStart.jsx
@@ -1,1 +1,0 @@
-//DESIGN START PAGE
